requirements.txt
# BASE
django==5.0.2
sqlparse==0.4.2
djangorestframework==3.14.0
django-filter==23.5
Pillow==10.2.0
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1 
Werkzeug==3.0.1
reportlab==4.2.0
openpyxl==3.1.2

# DEBUG
django-debug-toolbar==3.7.0

# DATABASE
mysql-connector-python==8.0.33
mysqlclient==2.1.1

# TEST
pytest-django==4.5.2
model_bakery==1.7.0
pytest-mock==3.8.2
pytest-cov==3.0.0

# NOTIFICATIONS
django-notifications-hq==1.7.0
pytest~=7.2.0

# CORS
django-cors-headers==3.13.0
whitenoise==6.6.0